+++ src/index.js
@@ -1,13 +1,12 @@
 import "./style.css";
 import Popup from "./js/Popup";
-import Api from "./js/Api"
-const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort5' : 'https://praktikum.tk/cohort5'
+import Api from "./js/Api";
+import CardList from "./js/CardList";
+import placesList from "./js/PL";
+const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort5' : 'https://praktikum.tk/cohort5';
 
-const placesList = document.querySelector('.places-list');
-const delButton = placesList.querySelector('.place-card__delete-icon');
 const popup = document.querySelector('.popup');
 const popupEdit = document.querySelector('.popup__edit');
-const popupImage = document.querySelector('.popup__image');
 const openButtonPopup = document.querySelector('.user-info__button')
 const form = document.forms.new;
 const name = form.elements.name;
@@ -19,83 +18,7 @@ const userPhoto = document.querySelector('.user-info__photo');
 const userName = document.querySelector('.user-info__name');
 const userJob = document.querySelector('.user-info__job');
 
-function openImagePopup(imageLink) {
-  imageLink = event.target.closest('.place-card__image');
-  if (!imageLink) return; 
-    if (!placesList.contains(imageLink)) return; 
-    imageLink=imageLink.getAttribute('style').slice(22, -1); 
-    const img = document.querySelector('.popup__content_img');
-    img.setAttribute('src', imageLink);
-    popupImg.open();
-}
-class Card {
-  constructor (name, link, openImageCallback) {
-    this.name = name;
-    this.link = link;
-    this.openImageCallback = openImageCallback;
-    this.placeCard = this.create(name, link);
-    this.like = this.like.bind(this);
-    this.remove = this.remove.bind(this);
-    this.openImage = this.openImage.bind(this);
-    this.placeCard.querySelector('.place-card__like-icon').addEventListener('click', this.like);
-    this.placeCard.querySelector('.place-card__delete-icon').addEventListener('click', this.remove);
-    this.placeCard.addEventListener('click', this.openImage);    
-  }
-
-  like () {
-    this.placeCard.querySelector('.place-card__like-icon').classList.toggle('place-card__like-icon_liked');
-  }
-
-  remove () {   
-    this.placeCard.parentNode.removeChild(this.placeCard);
-  }
- 
-  create () {
-    const placeCard = document.createElement('div')
-    placeCard.classList.add('place-card');
-
-    const cardImage = document.createElement('div'); 
-    cardImage.classList.add('place-card__image');
-    placeCard.appendChild(cardImage);
-    cardImage.setAttribute('style','background-image: url('+ this.link +')');
-  
-    const cardDescription = document.createElement('div');
-    cardDescription.classList.add('place-card__description');
-    placeCard.appendChild(cardDescription);
-  
-    const cardName = document.createElement('h3');
-    cardName.classList.add('place-card__name');
-    cardName.textContent = this.name;
-    cardDescription.appendChild(cardName);
-
-    let likeButton = document.createElement('button');
-    likeButton.classList.add('place-card__like-icon');
-    cardDescription.appendChild(likeButton);
-
-    let delButton = document.createElement('button');
-    delButton.classList.add('place-card__delete-icon');
-    cardImage.appendChild(delButton);
-  
-    return placeCard;
-  }
-
-  openImage() {
-    this.openImageCallback(this.link);
-  }
-}
-
-class CardList {
-  constructor(container, arr) {
-      this.container = container;
-      this.arr = arr;
-  }
-  addCard(name, link) {
-    const cardElement = new Card (name, link, openImagePopup);
-    this.container.appendChild(cardElement.placeCard);
-  }
-}
 const cardList = new CardList(placesList);
-
 
 const popupForm = new Popup (popup, 'popup_is-opened')
 openButtonPopup.addEventListener('click', function () {
@@ -105,7 +28,6 @@ const popupEditt = new Popup (popupEdit, 'open-edit')
 document.querySelector('.user-info__edit').addEventListener('click',function () {
   popupEditt.open();
 });
-const popupImg = new Popup (popupImage, 'open-image')
 
 const api = new Api({
   baseUrl: serverUrl,
