+++ src/js/PL.js
@@ -1,0 +1,2 @@
+const placesList = document.querySelector('.places-list');
+export default placesList;
